requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
idna==3.4
peewee==3.15.3
PyMySQL==1.0.2
python-dotenv==0.21.0
requests==2.28.1
urllib3==1.26.12
